Modificado ListaPreciosController para gestionar los items de la lista

+++ app/Http/Controllers/ListaPreciosController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\ListaPrecioItem;
 use App\ListaPrecios;
 use Illuminate\Http\Request;
 
@@ -14,7 +15,7 @@ class ListaPreciosController extends Controller
      */
     public function index()
     {
-        return response()->json(ListaPrecios::all());
+        return response()->json(ListaPrecios::with('listaPrecioItem')->get());
     }
 
 
@@ -52,8 +53,24 @@ class ListaPreciosController extends Controller
      */
     public function update(Request $request, ListaPrecios $listaprecio)
     {
-        $listaprecio->fill($request->json()->all())->save();
-        return response()->json($listaprecio);
+        $data = json_decode(file_get_contents('php://input'), true);
+
+        $listaprecio->fill($data)->save();
+        $listaprecio->load('listaPrecioItem');
+
+        $viejosItems = $listaprecio['listaPrecioItem'];
+        $nuevosItems = $data['lista_precio_item'];
+
+        //Se borran los items que no vinieron en el request
+        $itemsABorrar = $viejosItems->whereNotIn('id', array_column($nuevosItems, 'id'))->all();
+        ListaPrecioItem::destroy(array_column($itemsABorrar, 'id'));
+
+        //Se actualiza o se crea un item según si tiene seteado o no el id
+        foreach ($nuevosItems as $item){
+            $listaprecio->listaPrecioItem()->updateOrCreate(['id'=>(isset($item['id'])? $item['id'] : 0)], $item);
+        }
+
+        return response()->json($listaprecio->load('listaPrecioItem'));
     }
 
     /**
@@ -64,6 +81,9 @@ class ListaPreciosController extends Controller
      */
     public function destroy(ListaPrecios $listaprecio)
     {
+        if($listaprecio->listaPrecioItem()->count() !== 0){
+            return response()->json(['error' => 'No se ha podido eliminar la lista de precios ' . $listaprecio->nombre . ' porque existen ítems asociados'], 200);
+        };
         $listaprecio->delete();
         return response()->json('ok', 200);
     }
